Compute absolute module load order in build script
This supercedes the concept of module "layers" by providing a
definitive, hardcoded load order to ensure dependencies are satisfied
in order. This also removes the need for module dependencies to be
declared at runtime. Finally, this guarantees that static modules will
not form a cyclical dependency graph.

Some additional cleanup is still needed, and per-module dependency lists
need to be emitted to allow automatic loading of dependencies from an
incomplete request list.

Also, topographical sorting still needs to be implemented at runtime for
dynamic modules including a mechanism for detecting cyclical
dependencies.