Ovehraul how modules are configured by build script.
Instead of a separate CMakeLists.txt file for each module, we now parse
a .properties file for each path said to contain a module which contains
everything needed to configure most modules. The exception is
render_opengl since it invokes Aglet, so we also allow a module.cmake
file to be provided to allow for more complex per-module configuration.