lowlevel: Rename _ARGUS_ASSERT and replace some usages with assert
This macro was previously used for both assertions and checking preconditions,
which caused some ambiguity in the semantic meaning of calls to it. An assertion
should in theory never fail in production code, and if it does it is the result
of a faulty assumption, an invalid program state, or an incorrect logic flow.

On the other hand, we were mostly using _ARGUS_ASSERT to validate preconditions,
which _is_ permitted to faild in normal operation if (for instance) the library
is used incorrectly from client code or some constraint is exceeded. Given that
these constitute the majority of usages of the macro, it seems appropriate to
rename it to affirm_precond (where "affirm" is intended as a somewhat weaker
variation on "assert").

Most references were updated to point to the new name, although a few were
changed to the standard assert macro where they actually are making absolute
claims as to what should be possible. Note that this has been slightly stretched
to also include verifications that a library is behaving in spec, such as OpenGL
calls which are guaranteed to return positive integers.