test/lowlevel: Use GENERATE_COPY in Vector2 test
MSVC diverges from the C++ standard by requiring const/constexpr
variables to be explicitly captured. The C++ standard only requires
this if the variable is odr-used (which a constexpr cannot be), so this
is just another instance of MSVC doing whatever the hell it wants and
leaving us no choice but to work around it.