scripting/scripting_lua: Initial implementation for vector handling
This is very much a first draft: the code is still pretty messy and
will likely undergo significant refactoring in future revisions. Also,
vector reference parameters are not yet supported because the current
framework doesn't support it very well. We assume that there's always
a concrete instance of structs _somewhere_ that we can point to, but
this doesn't hold true for vectors passed by value since they're
not represented as C++ vectors (e.g. in Lua they're represented as
Lua tables).