scripting: Handle bound field types which aren't AutoCleanupable
This adds new behavior for bound fields which don't derive from
AutoCleanupable. Normally field values are returned as references,
which allows them to be mutated in-place if they are a class type.
However, in the case that a field's type isn't AutoCleanupable,
it will instead be returned as a const copy since we can't ensure
safety in returning a reference to it.