wm: Perform window reaping in separate loop instead of in Window::update
This moves the logic for cleaning up windows to a render thread loop
instead of performing it directly in Window::update. The old logic
had the window deleting itself which was causing all sorts of problems,
mainly to do with race conditions involving the window callback function.
The callback is now unregistered in the reaper loop as well.