scripting_lua: Fix lifetime issue in LuaCallback objects
We pass transitively ownership of LuaCallback objects via the closure
they're captured by to the C++ function which accepts them. This becomes
an issue when the C++ function passes the ownership to a structure which
outlive the lua_State, as when they're destroyed they attempt to clean
up after themselves even though the lua_State object has been freed.
This change sidesteps that issue by wrapping the lua_State in a C++
object and then passing a weak_ptr to it to the closure, so that we can
simply check whether the weak_ptr is valid before attempting to do
anything with it.