buildscript: Implement new 'aux' module type
This commit introduces a new module type for auxiliary modules.
Auxiliary modules are not built by CMake itself, but they still
contain a module.properties file which specifies dependencies and
can also be depended upon by other modules. This enables their
dependencies to be transitive such that if static module A depends
upon aux module B, and aux module B depends on static module C, a
dependency will implicitly exist from A to C which will then be
reflected in the generated module initialization order.

The intended initial use case for this is to allow for modules
which provide a Rust ABI for non-Rust modules as a crate. Thus,
a pure Rust module can depend on foo_rustabi and implicitly get
a dependency on foo.