core_rustabi: Rework module structure for C ABI bindings
This somewhat convoluted structure is designed to allow reuse of
generated bindings for dependency modules. i.e. The *_cabi module
can provide a use declaration for a dependency's bindings, and
then the `bindings` submodule can simply pull in that use
declaration from its parent. Making the bindings module nested
allows the declaration to be local to the C ABI module rather than
to the whole crate.