wm/render_*: Don't create contexts for temporary hidden windows
Previously the render backend would see the create event and try
to initialize a context, only for it to fail because the window
was already in the process of being destroyed. Context creation
will now be skipped entirely if the window has already received
a close request by the time the create event is propagated.