lowlevel: Improve vararg handling in logging functions
It's technically undefined behavior to use a non-POD type in the
last parameter before varargs begin, so we use a simple wrapper to
allow the corresponding C string to be passed to va_start instead.
This only raises warnings with MSVC, but I suspect this was
causing issues when logging crashes on Win32 (but have yet to
confirm this).