resman: Move static functions in C ABI outside of extern "C" block
Apparently it's technically UB to declare a static function as
extern "C" (if I understand correctly) which is effectively what is
being done by declaring it inside an extern "C" block. Clang and
GCC have no problem with this scenario, but MSVC throws a fit and
bails out if it sees an STL type in a function signature in a
declaration such as this.