scripting_rs: Add basic support for binding symbols from Rust
Most of the code introduced here is meant as a stopgap during the
transition from C++ to Rust and as such it is highly incomplete,
not fully tested, and honestly not very good. The only purpose is
to enable just enough scripting support to allow for simple test
scripts.

Note that memory safety goes completely out the window in bound
functions. For functions that return a reference, the return value
will be marked as mutable in the binding defintion and will be
coerced to a mutable reference when necessary (e.g. when modifying
fields). Eventually there will be a requirement that functions
return an Arc<Mutex<_>> or something like that if the return value
is supposed to be mutable.

Currently associated functions are not supported at all. This
actually _is_ needed for the stopgap so support will be implemented
soon.