buildscript: Link Rust libs together via Cargo
This fixes issues with dependency code and symbols being duplicated
across the static modules generated for Rust modules. In a
Rust-aware context, each dependency is only included a single time
when compiling into a single static library.

In particular, this fixes an issue with script bindings in
dependency crates being submitted multiple times, once by the
crate itself and once for each crate which depends on it.