game2d: Implement collision detection
This uses a naive check of the distance between the two objects squared
against the sum of their radii squared to cull well-separated objects.

Then the Minkowski difference between the two objects is computed by
walking along the convex hull of the resulting polygon and a second
culling check is done to return early if the polygon cannot possibly
contain the origin.

Finally, the moment of collision with each edge is computed and if any
fall within the last movement step, the motion is scaled appropriately
and the remaining motion is projected onto the colliding edge whereupon
another collision check is initiated. This last step enables "sliding"
alongside a surface while discarding the component of motion that is
perpendicular to it.